+++ src/react/components/modules/Register/styles.js
@@ -1,18 +1,17 @@
 import { createUseStyles } from 'react-jss'
 
 export default createUseStyles(theme => {
-  const primary = theme.color.primary
+  // console.log("Register-Theme: ", theme)
   return {
     root: {
-      height: '100%',
-      width: '100%',
-      backgroundColor: '#000',
-      color: primary,
+
     },
     button: {
       marginBottom: 10,
       width: 200,
       padding: 4,
     },
+    inputClass: {
+    }
   }
-}, { name: 'SignOnPage-styles' })
+}, {name: "Register-styles"})