+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### [2022-06-16]
+### Added
+- signon buttons, registration component, found with tsx/jss styles, struggled with thinking of okay-enough UI
+
 ### [2022-06-15]
 ### Added
 - react router, loginAuthProvider, pageTimeout, useAfterFirstMount